+++ src/icons/bear.tsx
@@ -69,3 +69,92 @@ export const MountainIcon: IconComponent = (props) => (
     <path d="M8 20l4-8 4 8" />
   </Icon>
 );
+
+export const AnvilIcon: IconComponent = (props) => (
+  <Icon {...props}>
+    <path d="M5 8h14l-1 3H6z" />
+    <path d="M8 11v6h8v-6" />
+    <path d="M6 20h12" />
+    <path d="M4 8c0-2 2-4 5-4h2" />
+  </Icon>
+);
+
+export const KilnIcon: IconComponent = (props) => (
+  <Icon {...props}>
+    <path d="M8 21V10l4-6 4 6v11" />
+    <path d="M8 14h8" />
+    <path d="M10 18h4" />
+  </Icon>
+);
+
+export const HarborIcon: IconComponent = (props) => (
+  <Icon {...props}>
+    <path d="M3 16c2 2 4 3 9 3s7-1 9-3" />
+    <path d="M5 13c1.5 1.2 3.5 2 7 2s5.5-.8 7-2" />
+    <path d="M12 4v9" />
+    <path d="M9 7h6" />
+  </Icon>
+);
+
+export const TorchIcon: IconComponent = (props) => (
+  <Icon {...props}>
+    <path d="M12 3c2 2 3 4 3 6a3 3 0 1 1-6 0c0-2 1-4 3-6z" />
+    <path d="M10 12h4v9h-4z" />
+  </Icon>
+);
+
+export const RailIcon: IconComponent = (props) => (
+  <Icon {...props}>
+    <path d="M4 8h16" />
+    <path d="M4 16h16" />
+    <path d="M8 4v16M16 4v16" />
+  </Icon>
+);
+
+export const InkIcon: IconComponent = (props) => (
+  <Icon {...props}>
+    <path d="M14 4l6 6-9 9H5v-6z" />
+    <path d="M13 5l6 6" />
+    <path d="M5 20h6" />
+  </Icon>
+);
+
+export const LingoIcon: IconComponent = (props) => (
+  <Icon {...props}>
+    <circle cx="12" cy="12" r="9" />
+    <path d="M3 12h18" />
+    <path d="M12 3c2.5 3 4 6 4 9s-1.5 6-4 9c-2.5-3-4-6-4-9s1.5-6 4-9z" />
+  </Icon>
+);
+
+export const SynapseIcon: IconComponent = (props) => (
+  <Icon {...props}>
+    <circle cx="6" cy="12" r="2" />
+    <circle cx="18" cy="6" r="2" />
+    <circle cx="18" cy="18" r="2" />
+    <path d="M8 12h6M16 7.5l-4 3M16 16.5l-4-3" />
+  </Icon>
+);
+
+export const WardenIcon: IconComponent = (props) => (
+  <Icon {...props}>
+    <path d="M12 3l8 4v5c0 5-3.5 8-8 9-4.5-1-8-4-8-9V7z" />
+    <path d="M9 12l2 2 4-4" />
+  </Icon>
+);
+
+export const CrucibleIcon: IconComponent = (props) => (
+  <Icon {...props}>
+    <path d="M8 6h8l-1 10H9z" />
+    <path d="M7 20h10" />
+    <path d="M12 3v3" />
+  </Icon>
+);
+
+export const BifrostIcon: IconComponent = (props) => (
+  <Icon {...props}>
+    <path d="M3 18c4-8 14-8 18 0" />
+    <path d="M5 18c3.5-6 10.5-6 14 0" />
+    <path d="M7 18c2.5-4 7.5-4 10 0" />
+  </Icon>
+);